[#1367] Let operator classes extend base operators
