+++ gradoop-flink/src/main/java/org/gradoop/flink/model/api/epgm/BaseGraphCollectionOperators.java
@@ -23,7 +23,14 @@ import org.gradoop.common.model.api.entities.Vertex;
 import org.gradoop.common.model.impl.id.GradoopId;
 import org.gradoop.common.model.impl.id.GradoopIdSet;
 import org.gradoop.flink.model.api.functions.GraphHeadReduceFunction;
-import org.gradoop.flink.model.api.operators.BaseGraphCollectionOperatorSupport;
+import org.gradoop.flink.model.api.operators.ApplicableUnaryBaseGraphToBaseGraphOperator;
+import org.gradoop.flink.model.api.operators.BinaryBaseGraphCollectionToBaseGraphCollectionOperator;
+import org.gradoop.flink.model.api.operators.BinaryBaseGraphCollectionToValueOperator;
+import org.gradoop.flink.model.api.operators.ReducibleBinaryBaseGraphToBaseGraphOperator;
+import org.gradoop.flink.model.api.operators.UnaryBaseGraphCollectionToBaseGraphCollectionOperator;
+import org.gradoop.flink.model.api.operators.UnaryBaseGraphCollectionToBaseGraphOperator;
+import org.gradoop.flink.model.api.operators.UnaryBaseGraphCollectionToValueOperator;
+import org.gradoop.flink.model.impl.operators.combination.Combination;
 import org.gradoop.flink.model.impl.operators.difference.Difference;
 import org.gradoop.flink.model.impl.operators.difference.DifferenceBroadcast;
 import org.gradoop.flink.model.impl.operators.distinction.DistinctById;
@@ -31,6 +38,7 @@ import org.gradoop.flink.model.impl.operators.distinction.DistinctByIsomorphism;
 import org.gradoop.flink.model.impl.operators.distinction.GroupByIsomorphism;
 import org.gradoop.flink.model.impl.operators.equality.CollectionEquality;
 import org.gradoop.flink.model.impl.operators.equality.CollectionEqualityByGraphIds;
+import org.gradoop.flink.model.impl.operators.exclusion.Exclusion;
 import org.gradoop.flink.model.impl.operators.intersection.Intersection;
 import org.gradoop.flink.model.impl.operators.intersection.IntersectionBroadcast;
 import org.gradoop.flink.model.impl.operators.limit.Limit;
@@ -38,6 +46,7 @@ import org.gradoop.flink.model.impl.operators.matching.transactional.Transaction
 import org.gradoop.flink.model.impl.operators.matching.transactional.algorithm.DepthSearchMatching;
 import org.gradoop.flink.model.impl.operators.matching.transactional.algorithm.PatternMatchingAlgorithm;
 import org.gradoop.flink.model.impl.operators.matching.transactional.function.AddMatchesToProperties;
+import org.gradoop.flink.model.impl.operators.overlap.Overlap;
 import org.gradoop.flink.model.impl.operators.selection.Selection;
 import org.gradoop.flink.model.impl.operators.tostring.functions.EdgeToDataString;
 import org.gradoop.flink.model.impl.operators.tostring.functions.EdgeToIdString;
@@ -63,7 +72,7 @@ public interface BaseGraphCollectionOperators<
   V extends Vertex,
   E extends Edge,
   LG extends BaseGraph<G, V, E, LG, GC>,
-  GC extends BaseGraphCollection<G, V, E, LG, GC>> extends BaseGraphCollectionOperatorSupport<LG, GC> {
+  GC extends BaseGraphCollection<G, V, E, LG, GC>> {
 
   //----------------------------------------------------------------------------
   // Base Graph / Graph Head Getters
@@ -319,5 +328,88 @@ public interface BaseGraphCollectionOperators<
    */
   default GC groupByIsomorphism(GraphHeadReduceFunction<G> reduceFunction) {
     return callForCollection(new GroupByIsomorphism<>(reduceFunction));
+  }
+
+  //----------------------------------------------------------------------------
+  // Call for Operators
+  //----------------------------------------------------------------------------
+
+  /**
+   * Creates a value using the given unary collection to value operator.
+   *
+   * @param operator unary graph collection to value operator
+   * @param <T> return type
+   * @return result of given operator
+   */
+  <T> T callForValue(UnaryBaseGraphCollectionToValueOperator<GC, T> operator);
+
+  /**
+   * Calls the given binary collection to value operator using this graph collection and the
+   * input graph collection.
+   *
+   * @param operator        binary collection to value operator
+   * @param otherCollection second input collection for operator
+   * @param <T> return type
+   * @return result of given operator
+   */
+  <T> T callForValue(BinaryBaseGraphCollectionToValueOperator<GC, T> operator, GC otherCollection);
+
+  /**
+   * Creates a graph collection using the given unary graph collection operator.
+   *
+   * @param operator unary graph collection to graph collection operator
+   * @return result of given operator
+   */
+  default GC callForCollection(UnaryBaseGraphCollectionToBaseGraphCollectionOperator<GC> operator) {
+    return callForValue(operator);
+  }
+
+  /**
+   * Calls the given binary collection to collection operator using this graph collection and the
+   * input graph collection.
+   *
+   * @param operator        binary collection to collection operator
+   * @param otherCollection second input collection for operator
+   * @return result of given operator
+   */
+  default GC callForCollection(BinaryBaseGraphCollectionToBaseGraphCollectionOperator<GC> operator,
+                               GC otherCollection) {
+    return callForValue(operator, otherCollection);
+  }
+
+  /**
+   * Applies a given unary graph to graph operator (e.g., aggregate) on each
+   * base graph in the graph collection.
+   *
+   * @param operator applicable unary graph to graph operator
+   * @return collection with resulting logical graphs
+   */
+  default GC apply(ApplicableUnaryBaseGraphToBaseGraphOperator<GC> operator) {
+    return callForValue(operator);
+  }
+
+  /**
+   * Transforms a graph collection into a graph by applying a
+   * {@link ReducibleBinaryBaseGraphToBaseGraphOperator} pairwise on the elements of the collection.
+   *
+   * @param operator reducible binary graph to graph operator
+   * @return base graph returned by the operator
+   *
+   * @see Exclusion
+   * @see Overlap
+   * @see Combination
+   */
+  default LG reduce(ReducibleBinaryBaseGraphToBaseGraphOperator<GC, LG> operator) {
+    return callForValue(operator);
+  }
+
+  /**
+   * Calls the given unary collection to graph operator for the collection.
+   *
+   * @param operator unary collection to graph operator
+   * @return result of given operator
+   */
+  default LG callForGraph(UnaryBaseGraphCollectionToBaseGraphOperator<GC, LG> operator) {
+    return callForValue(operator);
   }
 }
