+++ gradoop-flink/src/main/java/org/gradoop/flink/model/api/epgm/GraphCollectionOperators.java
@@ -16,8 +16,10 @@
 package org.gradoop.flink.model.api.epgm;
 
 import org.apache.flink.api.java.DataSet;
+import org.gradoop.common.model.impl.pojo.EPGMEdge;
+import org.gradoop.common.model.impl.pojo.EPGMGraphHead;
+import org.gradoop.common.model.impl.pojo.EPGMVertex;
 import org.gradoop.flink.io.api.DataSink;
-import org.gradoop.flink.model.api.operators.BaseGraphCollectionOperatorSupport;
 import org.gradoop.flink.model.impl.epgm.GraphCollection;
 import org.gradoop.flink.model.impl.epgm.LogicalGraph;
 
@@ -26,7 +28,8 @@ import java.io.IOException;
 /**
  * Defines the operators that are available on a {@link GraphCollection}.
  */
-public interface GraphCollectionOperators extends BaseGraphCollectionOperatorSupport<LogicalGraph, GraphCollection> {
+public interface GraphCollectionOperators extends BaseGraphCollectionOperators<EPGMGraphHead, EPGMVertex,
+  EPGMEdge, LogicalGraph, GraphCollection> {
 
   //----------------------------------------------------------------------------
   // Auxiliary operators
